Refactored the render utils into a separate file
