+++ src/modules/renderutils.js
@@ -1,0 +1,77 @@
+import {
+    Vector2D
+} from './vector2d.js';
+
+class RenderBuffer {
+    constructor(canvas) {
+        this.canvas = canvas;
+        this.context = canvas.getContext("2d");
+        this.width = canvas.width;
+        this.height = canvas.height;
+        this.stride = 4;
+
+        this.imagedata = this.context.createImageData(canvas.width, canvas.height);
+    }
+
+    plotPixel(x, y, color) {
+        const pixelIndex = ((y * this.imagedata.width) + x) * this.stride;
+
+        this.imagedata.data[pixelIndex] = color.r;
+        this.imagedata.data[pixelIndex + 1] = color.g;
+        this.imagedata.data[pixelIndex + 2] = color.b;
+        this.imagedata.data[pixelIndex + 3] = color.a;
+    }
+
+
+    applyImageData = () => {
+        this.context.putImageData(this.imagedata, 0, 0);
+    }
+}
+
+// Rendering functions
+function drawRect(startPos, endPos, color, renderBuffer) {
+    for (let x = startPos.x; x <= endPos.x; x++) {
+        for (let y = startPos.y; y <= endPos.y; y++) {
+            renderBuffer.plotPixel(x, y, color);
+        }
+    }
+}
+
+function drawLineDDA(startPos, endPos, color, renderBuffer) {
+    const dx = endPos.x - startPos.x;
+    const dy = endPos.y - startPos.y;
+
+    let steps;
+    if (Math.abs(dx) > Math.abs(dy)) {
+        steps = Math.abs(dx);
+    } else {
+        steps = Math.abs(dy);
+    }
+
+    const xIncrement = dx / steps;
+    const yIncrement = dy / steps;
+
+    let x = startPos.x;
+    let y = startPos.y;
+
+    //plot initial pixel
+    renderBuffer.plotPixel(Math.round(x), Math.round(y), color);
+
+    for (let stepIndex = 0; stepIndex < steps; stepIndex++) {
+        x += xIncrement;
+        y += yIncrement;
+
+        renderBuffer.plotPixel(Math.round(x), Math.round(y), color);
+    }
+}
+
+function renderWallSegment(xPos, yStart, yEnd, color, renderBuffer) {
+    drawLineDDA(new Vector2D(xPos, yStart), new Vector2D(xPos, yEnd), color, renderBuffer);
+}
+
+export {
+    RenderBuffer,
+    drawRect,
+    drawLineDDA,
+    renderWallSegment
+}
